Added User ID and API Key validation checks to save settings

+++ includes/class-eio-settings-manager.php
@@ -108,8 +108,61 @@ final class EIO_Settings_Manager {
 			$api_key = $api_key[count($api_key) - 1];
 		}
 		
+		if (false === is_string($user_guid) || empty($user_guid)) {
+			eio_add_error_notice(
+				__('You must provide a <strong>User ID</strong>', 'extractor-io')
+			);
+			
+			return;
+		}
+		
+		if (false === is_string($api_key) || empty($api_key)) {
+			eio_add_error_notice(
+				__('You must provide an <strong>API Key</strong>', 'extractor-io')
+			);
+			
+			return;
+		}
+		
+		$user_guid = strtolower(preg_replace('/\s+/', '', $user_guid));
+		$api_key = preg_replace('/\s+/', '', $api_key);
+		
+		if (false !== strpos($api_key, ':')) {
+			$api_key_components = explode(':', $api_key);
+			
+			if (1 !== count($api_key_components)) {
+				eio_add_error_notice(
+					__('The API Key is not formatted correctly.', 'extractor-io')
+				);
+				
+				return;
+			}
+		}
+		
+		try {
+			$import_io = new ImportIO($user_guid, $api_key);
+		} catch (BadFunctionCallException $e) {
+			eio_add_error_notice($e->getMessage());
+			
+			return;
+		}
+
+		$user = $import_io->current_user();
+		
+		if (false === is_array($user) || empty($user) || false === array_key_exists('username', $user) || false === is_string($user['username'])) {
+			eio_add_error_notice(
+				__('The User ID and or API Key were incorrect.', 'extractor-io')
+			);
+			
+			return;
+		}
+		
 		EIO()->options->update_option('user_guid', $user_guid);
 		EIO()->options->update_option('api_key', $api_key);
+
+		eio_add_updated_notice(
+			__('The <strong>User ID</strong> and <strong>API Key</strong> have been saved.', 'extractor-io')
+		);
 	}
 	
 	/**
@@ -123,6 +176,10 @@ final class EIO_Settings_Manager {
 	private function save_connector_mapping_settings() {
 		if (is_array($_POST['eio_import_to'])) {
 			EIO()->connector_mappings->update_option($_GET['connector'], $_POST['eio_import_to']);
+
+			eio_add_updated_notice(
+				__('The connector mappings have been saved.', 'extractor-io')
+			);
 		}
 	}
 }
